Refactored code to use async/await method, updated register function to include vulnerabilities, and structured API routes.
