--- DB/mysqldb.js
@@ -1,25 +1,0 @@
-const mysql = require('mysql2');
-const dotenv = require('dotenv');
-
-dotenv.config();
-
-const pool = mysql.createPool({
-  host: process.env.DB_HOST,
-  port: process.env.DB_PORT,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-  connectionLimit: process.env.CONNECTION_LIMIT || 10
-});
-
-// test the connection for error or success
-pool.getConnection((err, connection) => {
-  if (err) {
-    console.error('Error connecting to MySQL:', err.message);
-  } else {
-    console.log('Connected to the MySQL database successfully.');
-    connection.release();
-  }
-});
-
-module.exports = pool;
